Rewrote tree felling code.
Axes now have a maximum tree size that they are allowed to fell.
Added ability to repair stone tools by adding a loose rock to the crafting grid along with the tool. Repairing in this way lowers the max durability of the tool each time and only repairs half of the damage per craft.